+++ ui/src/editor/completion/RuleBasedPredictor.js
@@ -283,11 +283,11 @@ function withPostfix({ topHit }) {
 }
 
 // TODO: possibly a Jedi bug causing those keywords not showing up; remove if fixed in a future release
-function keywords({ input }) {
-    if ('True'.startsWith(input)) return 'True'
-    if ('False'.startsWith(input)) return 'False'
-    if ('None'.startsWith(input)) return 'None'
-}
+//function keywords({ input }) {
+//    if ('True'.startsWith(input)) return 'True'
+//    if ('False'.startsWith(input)) return 'False'
+//    if ('None'.startsWith(input)) return 'None'
+//}
 
 class RuleBasedPredictor {
     constructor(context) {
@@ -304,7 +304,7 @@ class RuleBasedPredictor {
             sequentialVariableNaming,
             forElementInCollection,
             suggestInitInsideClass,
-            keywords,
+            // keywords,
         ]
     }